refactor: relocate _walk_canonical to canonical/xpath_walker.py (run3 LOW)
run3 walk-canonical-vendor-leaf: the shared canonical-tree xpath walker
— the sole input to migration_validate.validate_against and the source
every codec's iter_xpaths yields from — lived inside the cisco_iosxe_cli
*vendor* codec, which the other 11 codecs + the honesty tests reached
into via `from ..cisco_iosxe_cli.codec import _walk_canonical`.

Move the function verbatim to a new
netcanon/migration/canonical/xpath_walker.py (a vendor-neutral home
alongside intent.py) and **re-export** it from cisco_iosxe_cli.codec so
every historical import path — the 11 lazy codec imports, the two
honesty tests, models/migration.py — keeps resolving unchanged. The
function body is byte-identical (extracted, not retyped).

No behaviour change: the walker output is identical, and the cross-mesh
matrix does not consume it (run_full_mesh compares model_dump() output
directly — noted in the walker's own docstring), so there is nothing to
regenerate. The live-validation honesty suite that DOES exercise it
(test_registry_capability_honesty, test_walk_canonical_coverage,
cisco_iosxe capability-matrix honesty) is green, as is the full
tests/unit + tests/integration gate.
